requirements.txt
lightlang[all]~=0.3.1
python-dotenv~=1.0.1
streamlit~=1.38
requests